requirements.txt
py-cord
python-dotenv
rich
SQLAlchemy
psycopg2-binary
PyMySQL
wavelink
requests
